% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_municipality.R
\name{check_municipality}
\alias{check_municipality}
\title{Check whether given municipality name is present in vg250 dataset}
\usage{
check_municipality(x = NULL)
}
\arguments{
\item{x}{character. Municipality name.}
}
\value{
logical.
}
\description{
Check whether given municipality name is present in vg250 dataset
}
\examples{
check_municipality(x = "Aachen")
check_municipality(x = "Roth")

check_municipality(x = "Aix la Chapelle")
check_municipality(x = "Freiburg")
}
