% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_municipality.R
\name{is_municipality}
\alias{is_municipality}
\title{Check whether given municipality name is present in vg250 dataset}
\usage{
is_municipality(x = NULL)
}
\arguments{
\item{x}{character. Municipality name.}
}
\value{
logical.
}
\description{
Check whether given municipality name is present in vg250 dataset
}
\examples{
\dontrun{
is_municipality("Aachen")
is_municipality("Aix la Chapelle")

is_municipality("Roth")
is_municipality("Freiburg")
}
}
