% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_present.R
\name{is_present}
\alias{is_present}
\title{Check whether given name of the administrative area is present in the dataset at the chosen level}
\usage{
is_present(x = NULL, level = "GEM")
}
\arguments{
\item{x}{character. Municipality name.}

\item{level}{character. Relevant administrative level.}
}
\value{
logical.
}
\description{
Check whether given name of the administrative area is present in the dataset at the chosen level
}
\examples{
# municipalities
is_present("Aix la Chapelle")
is_present("Aachen")

is_present("Roth")
is_present("Freiburg")

# districts
is_present("Aachen", level = "KRS")
is_present("Städteregion Aachen", level = "KRS")

# states
is_present("Westfalen", level = "LAN")
is_present("Nordrhein-Westfalen", level = "LAN")
}
