% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_centroid.R
\name{get_centroid}
\alias{get_centroid}
\title{Get centroid of the object queried from the dataset}
\usage{
get_centroid(x = NULL, level = "GEM")
}
\arguments{
\item{x}{character. Municipality name.}

\item{level}{character. Relevant administrative level.}
}
\value{
Geometry set of class \code{sfc_POINT}.
}
\description{
Get centroid of the object queried from the dataset
}
\examples{
get_centroid(x = "Aachen")
get_centroid(x = "Städteregion Aachen", level = "KRS")
}
