% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_centroid.R
\name{get_centroid}
\alias{get_centroid}
\title{Get centroid of the object queried from the dataset}
\usage{
get_centroid(x = NULL, level = "GEM", crs = "epsg:4326")
}
\arguments{
\item{x}{character. Name of administrative area.}

\item{level}{character. Relevant administrative level.}

\item{crs}{character. Coordinate reference system definition.}
}
\value{
Geometry set of class \code{sfc_POINT}.
}
\description{
Get centroid of the object queried from the dataset
}
\examples{
get_centroid("Aachen")
get_centroid("Aachen", crs = "epsg:25832")

get_centroid("Städteregion Aachen", level = "KRS")
}
