% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_centroid.R
\name{get_centroid}
\alias{get_centroid}
\title{Retrieve centroid of the object queried from the dataset}
\usage{
get_centroid(x = NULL)
}
\arguments{
\item{x}{character. Municipality name.}
}
\value{
Geometry set of class \code{sfc_POINT}.
}
\description{
Retrieve centroid of the object queried from the dataset
}
\examples{
get_centroid("Aachen")
}
