% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extent.R
\name{get_extent}
\alias{get_extent}
\title{Retrieve the bounding box information of the object queried from the dataset}
\usage{
get_extent(x = NULL, as_bbox = FALSE)
}
\arguments{
\item{x}{character. Municipality name.}

\item{as_bbox}{logical. Return bounding box as polygon or xmin/ymin/xmax/ymax?}
}
\value{
Geometry set of class \code{sfc_POLYGON} or extent of class \code{bbox}.
}
\description{
Retrieve the bounding box information of the object queried from the dataset
}
\examples{
get_extent("Aachen")
get_extent("Monschau", as_bbox = TRUE)
}
