% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geometry.R
\name{get_geometry}
\alias{get_geometry}
\title{Retrieve the original geometry information of the object queried from the dataset}
\usage{
get_geometry(x = NULL)
}
\arguments{
\item{x}{character. Municipality name.}
}
\value{
Geometry set of class \code{sfc_MULTIPOLYGON}.
}
\description{
Retrieve the original geometry information of the object queried from the dataset
}
\examples{
get_geometry("Aachen")
}
