% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geometry.R
\name{get_geometry}
\alias{get_geometry}
\title{Get original geometry information of the object queried from the dataset}
\usage{
get_geometry(x = NULL, level = "GEM", crs = "epsg:4326")
}
\arguments{
\item{x}{character. Name of administrative area.}

\item{level}{character. Relevant administrative level.}

\item{crs}{character. Coordinate reference system definition.}
}
\value{
Geometry set of class \code{sfc_MULTIPOLYGON}.
}
\description{
Get original geometry information of the object queried from the dataset
}
\examples{
get_geometry("Aachen")
get_geometry("Aachen", crs = "epsg:25832")

get_geometry("Städteregion Aachen", level = "KRS")
}
